Add GPU_FLAGS for session_manager
Add another set of flags (GPU_FLAGS) that are applied to session_manager.

Currently, we use the USE flag gpu_sandbox_allow_sysv_shm to pass
--gpu-sandbox-allow-sysv-shm to session_manager.  Other, similar USE flags
can be used to pass GPU-related flags.

The --gpu-sandbox-allow-sysv-shm parameter is used to allow the shmat system
call to be used by the Tegra GPU sandbox.

BUG=none
TEST=emerge-puppy chromeos-login, see that gpu_sandbox_allow_sysv_shm is set in
/etc/session_manager_use_flags.txt.

Change-Id: If83b154c96e00d3011b21e8a75ace44334ed8fdd
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/62380
Reviewed-by: Chris Masone <cmasone@chromium.org>