Add support for Chrome-gesture experiment
Previously, Salsa only supported experiments that worked by changing the
xinput parameters using the commandline tool in Crosh.  I've added a
similar tool for chromeOS located within the browser at chrome://salsa
now and it takes in the same encoded strings, but changes the values of
Chrome Gesture Preferences instead.

This patch adds the UI for selecting which type of experiment you want
to create when you're making new experiments and changes the
participant's page depending on which type it is.

eg: if you selected chrome-gestures when you created an experiment, your
participants will be given slightly different instructions.

BUG=chromium:249762
TEST=manually run with dev server on my local machine at
cmooney.mtv:8081

Change-Id: I8aecc3499b13adf835d24f0dbb7820bbfc02f2b0
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>