power: Remove FileTagger class.
This removes the FileTagger class, which touched and
unlinked files in /var/lib/power_manager to let
crash-reporter know whether unclean shutdowns occurred while
the system was suspended. It replaces it with much-simpler
code in the Daemon class.

The old code also tracked whether the system was in a
low-battery state or not, which I don't think makes sense
anymore: we shut down immediately when the battery is low.

BUG=chromium:259189
TEST=manual: cut battery power while system was suspended
     and checked that crash-reporter logged that it wasn't
     reporting the unclean shutdown; cut battery power while
     awake and checked that the unclean shutdown would've
     been reported

Change-Id: Ifdb3eb9c15a6eaaf36e1862659c55c4848b138ef
Reviewed-on: https://gerrit.chromium.org/gerrit/62550
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>