Fixed a bug in Platform::SyncPath when used with directories.
When using fsync with a file, opening a file descriptor using
open(O_WRONLY) is fine but it doesn't work with directories.  The
testing on the original CL was clearly not effective because attestation
enrollment fails.

To reproduce run platform_Attestation autotest on a system which has not
previously enrolled for attestation.

BUG=chromium:254028
TEST=unit, manual

Change-Id: Iadbfb6c1ab85f8f9f61b93512c16e5fe9fdef76e
Reviewed-on: https://gerrit.chromium.org/gerrit/62925
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>