Add new DBus calls that Store/Load Enrollment data across powerwashes.
This CL adds two new DBus calls to cryptohome that persists the DMToken
and device_id across powerwash that can be used by Chrome to persist
enrollment of a device across a powerwash. This data is passed and loaded
as a GArray which is a serialized protobuf containing this data.

BUG=chromium:241313
TEST=Unittests.

Change-Id: I2a25e16c4b5dc58b1e409a997b84870fd95c1a62
Reviewed-on: https://chromium-review.googlesource.com/63314
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>