Made C_Finalize safer when called from a signal-handler.
Previously, a WaitableEvent instance was used to signal threads waiting
on C_WaitForSlotEvent.  This raised a SIGABRT if C_Finalize was called
during a signal handler.  Now a simple slow poll is used with a side
effect that waiting threads may take a while to realize C_Finalize has
been called.

Note: It should not be considered safe to call C_Finalize from a signal
handler; don't do this.  But for those that do, this change should make
it safer than it was.

BUG=chromium:243147
TEST=unit, network_VPNConnect/control.l2tpipsec_psk

Change-Id: I386720a958173541c0620be28ed4c8f9f9b63568
Reviewed-on: https://gerrit.chromium.org/gerrit/63122
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>