Removed reading event desc metadata.
Event desc metadata is duplicate data in all perf data files so far.
Previously we tried to be smart about using either eventdesc or the other
data, but it is complicated enough that we just won't bother.
The new perf data file broke the old reader because the event ids are
neither unique nor consecutive (which was assumed previously).

BUG=none
TEST=FEATURES=test emerge-amd64-generic quipper

Change-Id: I64b6ddba703b3c72634150fdec034e1116d1d211
Signed-off-by: Rohin Shah <rohinmshah@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63396
Reviewed-by: Simon Que <sque@chromium.org>