client: Add unittests for PickUrlForId() in main.
In order to have unittests for the main functions the relevant code
was moved to a separated new file (peer_selector.*) and a new class
is introduced to fake the avahi ServiceFinder.

Also, two small changes are also included in the ServiceFinder
interface: The returned size_t value for AvailableFiles was dropped
and the GetPeersForFile method now returns a vector of const
pointers.

BUG=chromium:264384
TEST=client unittests are being executed and pass.

Change-Id: Id7b88044ba0ce7f3a998cf8b96b37911fd3ea31c
Reviewed-on: https://gerrit.chromium.org/gerrit/63667
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>