Adjust event size for modified command string
Sometimes the comm string will be modified (e.g. replaced with Md5sum)
to be longer than the original.  As such, it will not fit in the
original |header.size| bytes of memory.

This patch introduces a check for the data size vs the size listed in
the header.  If there is a difference, adjust the size listed in the
header.

BUG=None
TEST=unit tests pass

Change-Id: I08a4bec71e39bd177176497ef59e1c66777002c1
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63508
Reviewed-by: Rohin Shah <rohinmshah@chromium.org>