shill: Handles scan request in the middle of an ongoing scan.
When exploring the scanning code, it occurred to me that we didn't
handle this case.  This CL adds a check for kScanIdle to the 'Scan' call
(this check had been desired, anyway, to ensure a clean entrance to the
scanning process) and discards the incoming scan request.  This approach
ends up bailing out of a scan a little earlier than previous, similar,
checks but I don't believe losing a few scans that occur _immediately_
after another scan or connect will pose much of a problem.

This change necessitated altering the code executed when progressive
scan fails over to full scan (for errors and for no connection found).
It used to call 'Scan' directly but, now, posts ScanTask in order to
bypass the new check in 'Scan'.

Once the mechanism for checking 'scan-while-connecting' became based on
scan_state, the tests for NoScansWhileConnecting needed to be rewritten.
It was discovered that the progressive scan version of this test had
used full scan in places it shouldn't have so that test needed rewriting,
anyway.  In addition, a couple new tests were added specifically for the
scan-while-scanning case.

I also reworked WiFiMainTest.ProgressiveScanError to match more closely
the actual in-the-wild operation of the code.  This required removing a
Geolocation check which depended on a side-effect of the previously poor
test modeling.

While I was in there, I had to make ResetScanTimer virtual so that the
mock would work.

BUG=chromium:266413
TEST=unittest

Change-Id: I4030b95a3bc82c403440dc401190d40a54f7a486
Reviewed-on: https://chromium-review.googlesource.com/63645
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>