shill: add some more unit tests for ppp dongle support
Following on CL:63937, this commit adds unit tests for the
changes made to cellular_capability_universal, in CL:51356.

Note that I've changed the ON_CALL statements in CreateDBusPropertiesProxy
to EXPECT_CALL...AnyNumber. This is arguably the wrong place to set
expectations on the number of calls. However, given that
CreateDBusPropertiesProxy is called from a loop inside OnListBearersReply,
there's not a good way to set expectations from an individual test.

(In particular, the DBusPropertiesProxy objects are allocated and
forgotten by the test fixture before any of the test-specific code
can grab a handle to them.)

While there: remove GetActiveBearerProperties and GetInactiveBearerProperties,
as a) they're unused, and b) they might be confused with the new methods
mutable_active_bearer_properties and mutable_inactive_bearer_properties.

BUG=chromium:246826
TEST=new unit tests

Change-Id: I328be4866bd4736ed4d7481a8faa60437b265196
Reviewed-on: https://gerrit.chromium.org/gerrit/63996
Reviewed-by: Arman Uguray <armansito@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>