shill: OpenVPNDriver: Use double vector of strings for arguments
Restructure the code that creates command-line arguments for the
openvpn process.  Instead of building a flat vector of strings,
each logical option supplied to openvpn is itself a vector of
strings.  Thus, it is possible to output either a command line
argument list for OpenVPN (as is still done here) or a configuration
file which will contain each multi-word option on a separate line.

This structure also allows testing to be more rigorous since
expectations won't conflate the flag with its arguments.

BUG=chromium:217624
TEST=Unit tests, network_VPNConnect.openvpn_user_pass

Change-Id: I7d2973792372d43df6c3a3ebe3728debd09e1e68
Reviewed-on: https://gerrit.chromium.org/gerrit/64292
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>