server: Implement server-side metrics.
This patch implements the server side metrics of the P2P subsystem.
These metrics are mostly recorded in p2p-http-server, but due to
security restrictions the actual metric is sent to libmetrics from
p2p-server. To achieve this, a pipe is used to transmit from
p2p-http-server to p2p-server messages in the form of a struct with
two fields: message_type and value.

The metrics implemented here are the following:
 * P2P.Server.RequestResult
 * P2P.Server.ContentServedSuccessfullyMB
 * P2P.Server.ContentServedInterruptedMB
 * P2P.Server.RangeBeginPercentage
 * P2P.Server.DownloadSpeedKBps
 * P2P.Server.ClientCount
 * P2P.Server.FileCount

BUG=chromium:260486
TEST=unittests. Manual test on a chromebook shows metrics in chrome://histograms/

Change-Id: I0e1e95afca8ad2ce70e4aebeba26cb2d148ec357
Reviewed-on: https://gerrit.chromium.org/gerrit/65075
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>