http-server: Don't include time sleeping in transfer budget
If a client is requesting a resource for which we have to wait
(e.g. expected size is greater than actual size), the client gets an
unfair speed bump when we resume since we include the time spent
sleeping in the calculation to limit the download speed. Fix this by
disregarding the time spent sleeping when figuring out if we need to
sleep or not.

Also move the ClockInterface/Clock/FakeClock machinery from client
into common and use it. Whilst doing this, slightly rework
ClockInterface::Sleep to take a TimeDelta instead of an unsigned int.

BUG=None
TEST=New unit test + unit tests pass
Change-Id: I6c3440930c42b5e6567cc33129d0a61eb3a79435
Reviewed-on: https://gerrit.chromium.org/gerrit/64852
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: David Zeuthen <zeuthen@chromium.org>
Tested-by: David Zeuthen <zeuthen@chromium.org>