server: Replace timeouts by condition functions in unittests
Most of the unittests in the p2p-server and p2p-http-server programs
need to run the GLib main loop to attend events. The current
implementation was running the main loop for a fixed period of time
without checking if all the expected events were attended allowing
the test to finish earlier (and fail) or to take more time than it
actually needed.

This fix replaces all those timeout based main loop executions with
two stragies:
 * In the case where all the events involved in the test are already
   waiting to be executed from the main loop, the main loop is
   iterated while there is some work to do or a maximum number of
   iterations is reached.
 * In the case where externally generated events are not generated
   beforehand, a termination callback function is used to run the
   main loop. This provided callback should flag when is ok to stop
   the main loop execution. In this case, the program will be
   blocked waiting for the events to be generated or a biger timeout
   is reached. This timeout (1 minute by default) is used to stop
   the test when the expected events are not generated and supposedly
   will not be generated to make the test fail with the missed
   expectations.

Other minor cleanups to the unittests are included in this patch.

BUG=chromium:270427
TEST=unittests!

Change-Id: I13336e4a9924f4b36538a221279f99cb40231121
Reviewed-on: https://gerrit.chromium.org/gerrit/65797
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>