shill: dongles: UpdateBearerPath when modem enters Connected state
In order to fire up PPP for a dongle, shill needs to know that the
ModemManager Bearer requires PPP. To know that, shill needs to
call ListBearers on the appropriate mm1::ModemProxy.

In the normal case, where shill connects the Cellular modem, this
happens through CellularCapabilityUniversal::OnModemStateChangedSignal.

However, in the case where the Cellular modem is already connected
at the time shill starts (as happens if shill crashes after connecting
the modem), OnModemStateChangedSignal will not be reached. That's
because, from ModemManager's perspective, the modem stays in the
Connected state the whole time.

To resolve this, add a call to UpdateBearerPath in
CellularCapabilityUniversal::OnModemStateChanged (note that this method
is different from CellularCapabilityUniversal::OnModemStateChangedSignal.)
Since this method is invoked at the time shill learns of the modem,
we're now guaranteed that UpdateBearerPath will be called. And that
call will eventually lead to the capability calling StartPPP, if
appropriate.

BUG=chromium:261710
TEST=new unit tests, manual

Manual test
-----------
1. plug in ppp dongle
2. wait for network to connect
3. killall -9 shill
4. wait for network to connect
   (should happen in ~10 seconds)

Change-Id: I2ea1581dd33a9969332968ea7c160e669359b4f0
Reviewed-on: https://gerrit.chromium.org/gerrit/66032
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>